pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.29.0",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
